Create helpers.py and update tests and examples
