+++ examples/aliases/README.md
@@ -2,8 +2,11 @@
 
 
 ## Notes about `option_strings`:
-- Additional names for the same argument can be added via the `field`
-function.
+Additional names for the same argument can be added via the `alias` argument
+of the `field` function (see [the custom_args Example](
+  /examples/custom_args/README.md) for more info).
+
+The `simple_parsing.ArgumentParser` accepts an argument (currently called `add_option_string_dash_variants`, which defaults to False) which adds additional variants to allow using either dashes or underscores to refer to an argument:
 - Whenever the name of an attribute includes underscores ("_"), the same
 argument can be passed by using dashes ("-") instead. This also includes
 aliases.
@@ -28,7 +31,7 @@ class RunSettings:
     debug: bool = field(alias=["-d"], default=False)
     some_value: int = field(alias=["-v"], default=123)
 
-parser = ArgumentParser()
+parser = ArgumentParser(add_option_string_dash_variants=True)
 parser.add_arguments(RunSettings, dest="train")
 parser.add_arguments(RunSettings, dest="valid")
 parser.print_help()
