+++ examples/inheritance/inheritance_example.py
@@ -1,5 +1,5 @@
 from simple_parsing import ArgumentParser
-from simple_parsing.utils import JsonSerializable
+from simple_parsing.helpers import JsonSerializable
 
 
 from dataclasses import dataclass
