+++ simple_parsing/__init__.py
@@ -3,11 +3,12 @@
 """
 from . import utils, wrappers
 from .parsing import ArgumentParser, ConflictResolution
-from .utils import (Formatter, InconsistentArgumentError, MutableField, choice,
-                    field, subparsers)
+from .utils import (Formatter, InconsistentArgumentError)
+from .helpers import field, choice, MutableField
 
 __all__ = [
     "Formatter", "InconsistentArgumentError", "MutableField", "choice",
     "field", "subparsers",
     "ArgumentParser", "ConflictResolution",
+    "field", "choice", "MutableField"
 ]
