--- simple_parsing/wrappers/bob.py
@@ -1,8 +1,0 @@
-import dataclasses
-
-@dataclasses.dataclass
-class Bob:
-    a: int
-
-for field in dataclasses.fields(Bob):
-    print(field.default)
