+++ test/test_lists.py
@@ -7,7 +7,7 @@ import pytest
 
 import simple_parsing
 from simple_parsing import ArgumentParser, InconsistentArgumentError
-from simple_parsing.utils import list_field
+from simple_parsing.helpers import list_field
 from .testutils import *
